pyproject.toml
[project]
name = "ashby-sdk"
version = "0.4.0"
description = "Python SDK for the Ashby ATS API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Marvin Gabler" }
]
keywords = ["ashby", "ats", "api", "sdk", "recruiting", "hr", "applicant-tracking"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "responses>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/deepweather/ashby_python_sdk"
Documentation = "https://github.com/deepweather/ashby_python_sdk#readme"
Repository = "https://github.com/deepweather/ashby_python_sdk.git"
Issues = "https://github.com/deepweather/ashby_python_sdk/issues"
Changelog = "https://github.com/deepweather/ashby_python_sdk/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ashby_sdk"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.lint.isort]
known-first-party = ["ashby_sdk"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
strict = true
packages = ["ashby_sdk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ashby_sdk --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests requiring real API",
    "writes: marks tests that write/modify data in Ashby",
]
